fix: break circular dependent
